Moved all binding methods from Spec to Binding, made validator names same as method that calls them to allow mapping, and reworked spec JSON generation for new combined format with BoundSpec that includes conditions and validations as part of that.
